requirements.txt
aws-cdk-lib>=2.0.0, <=2.21.1
constructs>=10.0.0
types-boto3
